requirements.txt
hvac
hvac[parser]
requests
python-dotenv
dataclasses
dataclasses_json